// src/app/services/api-backend/model/municipio.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Sucursale } from './sucursale';
import { Departamento } from './departamento';


export interface Municipio { 
    codMunicipio?: number;
    codDepartamento?: number;
    nombre?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    codDepartamentoNavigation?: Departamento;
    sucursales?: Array<Sucursale> | null;
}