// src/app/services/api-backend/model/sucursale.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Municipio } from './municipio';
import { SucursalesUsuario } from './sucursalesUsuario';
import { Sala } from './sala';


export interface Sucursale { 
    codSucursal?: number;
    codMunicipio?: number;
    telefono?: string | null;
    direccion?: string | null;
    nombre?: string | null;
    activa?: string | null;
    usuarioIng?: string | null;
    fechaIng?: string;
    usuarioAct?: string | null;
    fechaAct?: string | null;
    codMunicipioNavigation?: Municipio;
    salas?: Array<Sala> | null;
    sucursalesUsuarios?: Array<SucursalesUsuario> | null;
}